+++ core-ng/src/test/java/core/framework/impl/log/LoggerImplTest.java
@@ -15,5 +15,7 @@ class LoggerImplTest {
         assertEquals("c.f.Bean", LoggerImpl.abbreviateLoggerName("core.framework.Bean"));
         assertEquals("c.f.a.Bean", LoggerImpl.abbreviateLoggerName("core.framework.api.Bean"));
         assertEquals("c.f.a.module.Bean", LoggerImpl.abbreviateLoggerName("core.framework.api.module.Bean"));
+        assertEquals("c.f.a.m.service.Bean", LoggerImpl.abbreviateLoggerName("core.framework.api.module.service.Bean"));
+        assertEquals("c.f.a.m.s.impl.Bean", LoggerImpl.abbreviateLoggerName("core.framework.api.module.service.impl.Bean"));
     }
 }